requirements.txt
customtkinter
pytesseract
Pillow
deep-translator
pymupdf
opencv-python
numpy
reportlab
supabase
python-dotenv
requests
deepl
openai
windnd
